Add method in the controller for the api request
